+++ train_control.lua
@@ -2,14 +2,33 @@
 -- Created By Jackz
 local SCRIPT = "train_control"
 local VERSION = "1.0.1"
--- Remove these lines if you want to disable update-checks: (6-11)
+local CHANGELOG_PATH = filesystem.stand_dir() .. "/Cache/changelog_" .. SCRIPT .. ".txt"
+-- Check for updates & auto-update: 
+-- Remove these lines if you want to disable update-checks & auto-updates: (7-54)
 util.async_http_get("jackz.me", "/stand/updatecheck.php?ucv=2&script=" .. SCRIPT .. "&v=" .. VERSION, function(result)
     chunks = {}
     for substring in string.gmatch(result, "%S+") do
         table.insert(chunks, substring)
     end
     if chunks[1] == "OUTDATED" then
-        util.toast(SCRIPT .. " has a new version available.\n" .. VERSION .. " -> " .. chunks[2] .. "\nDownload the latest version from https://jackz.me/sz")
+        -- Remove this block (lines 15-31) to disable auto updates
+        util.async_http_get("jackz.me", "/stand/changelog.php?raw=1&script=" .. SCRIPT .. "&since=" .. VERSION, function(result)
+            local file = io.open(CHANGELOG_PATH, "w")
+            io.output(file)
+            io.write(result:gsub("\r", "") .. "\n") -- have to strip out \r for some reason, or it makes two lines. ty windows
+            io.close(file)
+        end)
+        util.async_http_get("jackz.me", "/stand/lua/" .. SCRIPT .. ".lua", function(result)
+            local file = io.open(filesystem.scripts_dir() .. "/" .. SCRIPT .. ".lua", "w")
+            io.output(file)
+            io.write(result:gsub("\r", "") .. "\n") -- have to strip out \r for some reason, or it makes two lines. ty windows
+            io.close(file)
+
+            util.toast(SCRIPT .. " was automatically updated to V" .. chunks[2] .. "\nRestart script to load new update.", TOAST_ALL)
+        end, function(e)
+            util.toast(SCRIPT .. ": Failed to automatically update to V" .. chunks[2] .. ".\nPlease download latest update manually.\nhttps://jackz.me/stand/get-latest-zip", 2)
+            util.stop_script()
+        end)
     end
 end)
 
@@ -36,6 +55,15 @@ try_load_lib("natives-1627063482")
 
 while WaitingLibsDownload do
     util.yield()
+end
+-- Check if there is any changelogs (just auto-updated)
+if filesystem.exists(CHANGELOG_PATH) then
+    local file = io.open(CHANGELOG_PATH, "r")
+    io.input(file)
+    local text = io.read("*all")
+    util.toast("Changelog for " .. SCRIPT .. ": \n" .. text)
+    io.close(file)
+    os.remove(CHANGELOG_PATH)
 end
 
 local models = {